requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.0.0
python-dateutil>=2.8.0
numpy>=1.24.0
django-filter>=23.0
Pillow>=10.0.0
pytz>=2023.3
requests>=2.31.0 